+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 
 # dense
 
+
 ## dense 1.0.1  not yet released
+
+* Straighten Dense.fetch
+
 
 ## dense 1.0.0  released 2017-09-29